pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hsi-rgbd-calibkit"
version = "0.1.0"
description = "Calibration toolkit for rigid pushbroom HSI line-scan camera with OAK-D S2 RGB-D camera rig"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "HSI-RGBD Calibration Team"}
]
keywords = ["hyperspectral", "calibration", "rgb-d", "oak-d", "line-scan", "pushbroom"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "pyyaml>=6.0",
    "scipy>=1.7.0",
    "rich>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
hsi-rgbd-calib = "hsi_rgbd_calib.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/hsi-rgbd-calibkit"
Repository = "https://github.com/your-org/hsi-rgbd-calibkit.git"
Documentation = "https://github.com/your-org/hsi-rgbd-calibkit#readme"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
